Chain Indexing - Eth-Contracts Layer (#1031)
Asynchronously refresh peer connections and reflect the latest state of the protocol in the set of known 'peers' for the discovery provider during indexing. Why is this important? To ensure that any new nodes which have not yet been included in a replica set can still be indexed without issue - there is one particular scenario where a metadata or image file exists in a single location prior to being further distributed and there is a risk of indexing stoppage. This PR should fix that.