[AUD-1340] - TrackListens % of success vs failure (#2383)
Explicitly track the number of recorded listen tx submissions vs successes in redis for easier reporting.

As listens are received, a counter is generated for that particular hour in redis and incremented on submission and success. The difference between the two elements represents the number of transactions that failed to land.

We can leverage this data to understand when performance for writes is degrading.