Restart indexer on crash. Cleanup src.
Since pg client / pool can just emit errors
and it's not exactly clear the best way to handle in app code,
do the erlang thing and "let it crash" and use pm2 to restart,
because boot sequence is a known good state and will do a catchup.

The need to add pm2 to do this "inside" the existing container
having separate docker containers for stuff like this.