Notifications for milestones (track plays, reposts, etc)
Trigger captures old and new row so that listener can observe crossing of threshold.
However, while follow / repost notificaitons could be backfilled in catchup mode...
this milestones listener would have to be online to create notifications and wouldn't be able to "catchup"

I think a better long term strategy is to use triggers for all count / tally stuff,
and have the trigger create a milestone row in postgres when a threshold is crossed.
This approach would also reduce the "find work" overhead of current celery batch method.