simplify es-indexer startup behavior
before it would attempt to block server + celery start by running in
catchup mode, but if this failed it would prevent es-indexer from
starting and no indexer would run.

simply start es-indexer normally... if we want more coordination we can
do it in application code.  This is how it will be if es-indexer is in a
standalone container also.