User List pagination example.
it is kinda hard to generalize pagination on a nested collection...
so this instead uses a root level User query with a number of options.

the nice thing is that you can get fts.

initially was trying to use boost + scoring to do in one query,
but was tricky, so just used two separate queries:
one for "known" people I follow, one for unfollowed.

This is similar to how current client works.