Fix replicas for DN redis
An earlier fix for the assertRateLimiter tests exposed the Redis container on port 6379. But this breaks when swarming to create replicas, as the other replicas can't share that port. Instead, tests should be run inside the container so that they can connect to Redis, or we need to mock the Redis dependency in the test. This change makes the assumption of the former but it can change.