Fix local entity writes (#12594)
### Description
Two small issues:
1. We weren't correctly detecting dev environment for the relay plugin.
This was causing EM writes to fail because signer recovery was using the
wrong contract address.
2. The EM code in SDK was consuming the body of responses before
checking for errors. This makes the response body unusable in the case
that we throw a RequestError up the stack (since you can only consume it
once). The fetch logic in the generated `runtime` does it in the
opposite order to allow for calling code to do what it wants with the
response body in the error case.

NOTE: This will also fix e2e tests since they were failing on the track
upload test portion

### How Has This Been Tested?
`audius-cmd user create user1 && audius-cmd track upload --from user1`