[PE-6556] Adds trending playlists to explore page (#12587)
### Description
Implemented a lineup carousel that pulls the trending playlists lineup
and feeds that to `CollectionTiles`. This should make it so that viewing
the full trending playlists page keeps the playback state synchronized.

Also fixed playback issues with most of the sections using
`TrackTileCarousel` due to the uid formatting being strange.

Required a little bit of fussing with styles and types for
CollectionTile to fix some layout issues. I think I found all the places
where these are used and verified that they still look correct.

### How Has This Been Tested?
Local clients against staging