Make SDK mint agnostic and have solana-relay respect ArtistCoins table (#12603)
- Updates SDK to not internally manage fixed-decimal in the
`ClaimableTokensClient` or the `PaymentRouterClient` and instead require
the caller to pass the exact bigint amounts and derive authorities on
the fly
- Updates solana-relay to query the DB to get the artist coin mints and
use that instead of a hardcoded/configured list of mints for allowing
transactions through. Updates the tests to not have type errors. Mocks
the db call for the tests.
- Updates the storage pedalboard package with the latest db changes