Fix crash on AudioBreakdownModal (#12614)
My account was experiencing a crash on the `AudioBreakdownModal` because
we were mixing numbers and bigints in a reduce call. This brought my
attention to how `useWalletAudioBalances` wasn't returning the proper
types. After investigating other `useQueries` calls, I saw we're using a
wrapper and combiner, so I adopted that pattern as well.

Also fixes a warning about not having keys in WalletsTable