Fix staging ios build by avoiding harmony build (#12625)
### Description

There is a bug where we try to build harmony on an mac box and it
complains because it tries to build with darwin rollup, which we dont
have. The fix is to avoid building harmony for mobile, because we
technically dont need it. This requires going a bit manual on the
turbo.json side for mobile, which is a bit hacky. Lets use this to
unblock for now.