Make asset overview page responsive (#12636)
### Description
This PR makes the asset pages responsive as seen in the design. On
tablet and smaller screens, we move all of the content into a full width
single row column. The PR also introduces an error boundary state for
the `BalanceSection`

### How Has This Been Tested?

`npm run web:prod` and navigate to an asset page