[PE-6558][PE-6559][PE-6560] Empty search/explore states (#12639)
### Description
Apologies for the largeness. The web part of this got away from me due
to code reorg.
* Moved most sections into their own component files to make future perf
optimization re: delayed loading easier. The idea is I want each
component to fetch its own data and we were running a bunch of
simultaneous hooks in the top-level page on web.
* Cleaned up layout so the the outer container on all platforms is doing
the `gap` instead of using bottom margins in some places
* Fixed all of the hooks that supply data for these sections to properly
use `InifiniteQuery` and allow passing values for all of the SDK
parameters to them
* Note here: I took an approach of wrapping the underlying SDK get
params types with helpers that coerce the types into infinite query
style params (i.e. use of `pageSize` instead of `limit` and `offset`).
The goal is that we don't need to manually implement all of the
underlying SDK parameters, they are just transparently passed through
and show up in the types when using the hooks. Definitely open to
feedback on this pattern. It _could_ be turned into code generation or
something more generic, but I'm wary of too much magic here (there is
already a lot).
* Also **fun fact**, we were fetching 30 items for some of these queries
because that's the default...They are all set to 10 now and I made sure
that all usages of the hooks in search/explore explicitly pass values
for `timeRange` and `pageSize` so that we don't accidentally mess up
this page if we change those defaults in the future.
* Added boolean conditions for the various tabs to control which content
should show and updated the content sections for search/explore pages to
conditionally render the sections
* Removed the placeholder content, updated logic to either show search
results or the empty content for each tab.
* Fixed moods grid to update the current category with a mood instead of
hard-coding to 'tracks'

Going to do mobile web in a separate PR as it needs additional reorg

### How Has This Been Tested?
Lots of visual comparison of the clients against Figma :-)