[QA-2273] Fix type issue in useAudioBalance (#12652)
- https://github.com/AudiusProject/audius-protocol/pull/12617

#12617 changed the underlying cache object type without updating the
`getQueryKey` typing, which masked an issue in some optimism code. That
code would then try to add a bigint to an object, and then JS in it's
everlasting wisdom would autoconvert the object to a string and concat
the two. Then, when reading that object back, we think it's an object
with a `balance` property, but it's a string with no such property, and
is thus undefined. Unlike adding an object to a bigint, adding undefined
to a bigint throws an error (thanks JS) so that was the underlying root
cause of the issue, only present during swaps.

I updated the optimistic code to all be in one place and use the same
underlying function to prevent any future confusion.