[PE-6592] Update chat composer line indexing logic (#12662)
### Description
Bug with the composer when there were duplicate line text and index
combinations in the rendered text sections. Updated to update and track
the line number across the different text sections

Root cause of bug: There were multiple lines in the renderedTextSections
that had the same text value and index resulting in the same key, which
caused issues when trying to rerender and update them.

Fix: Added a lineIndex variable to keep track of the number of lines so
that the key will increment properly with the number of lines in the
rendered text section

### How Has This Been Tested?
Manually Tested