[QA-2255] Revert "[QA-2193] Fix android playback stuck on one track (#12460)" (#12677)
This reverts commit 4746f2ba9b1b1fac6f59af16e68405f8b23a08de.

### Description
[This PR](https://github.com/AudiusProject/audius-protocol/pull/12460)
caused a bug where if the user scrolled quickly on feed on mobile (both
ios and android), they could get into a state where the audio playback
would get stuck, or just play silence.

By reverting changes to `AudioPlayer.tsx` @dylanjeffers and i found that
this PR was the root cause.

We'll revert that commit and come back and try to solve the original bug
of that PR. That bug was less severe, android-only and only on
collections.

### How Has This Been Tested?

Tested manually on local ios prod sim