Add address lookup table to minimize swap size (#12685)
### Description
Added a new address lookup table to the `useSwapTokens` hook in order to
help reduce swap byte usage. Also added an arbitrary `20` max accounts
to `Jupiter.ts` to minimize the amount of hops needed in a swap. We
should monitor swap failure rate and adjust as needed.

### How Has This Been Tested?

`npm run web:prod` 

and swap between bonk and audio a few times