[PE-6626][PE-6580][PE-6581] Fix slow loading when switching explore tabs on web (#12687)
### Description
I tried a few things and landed here.
Using `display: none` and keeping things mounted didn't noticeably
improve anything because the majority of the slowness is due to having
to re-render a _lot_ of elements whenever we change pages.
It turns out to be faster to just let them unmount and then adopt the
following strategy:
* Use an intersection observer to determine if a given section is within
a few hundred pixels of the bottom of the screen
* Disable the data fetching hook until the observer fires
* Render loading skeletons by default until the observer fires
* Use the `triggerOnce` mode so that `inView` stays true after the first
time. The slowness isn't with having all the elements on the page at
once, it's trying to render them all together the first time (including
their data fetching hooks).

This effectively breaks up the rendering enough that switching tabs and
the initial load are much faster.

There are still some other things that could be addressed to make this
snappier (virtualization of the nested lists for example). I will follow
up in another PR to do those. This was getting a little large.

Some more details on changes:
* Added a hook that wraps `useInView` and passes some hard-coded options
to make things consistent. I may update this thing in the future to have
a little more functionality, tbd.
* We're hard-coded to 300px margin (so things load when they are within
300px of being scrolled into view), no lower threshold on how much of
the component needs to be visible
* We use the main content ref as a viewport target on desktop, device
viewport on mobile
* We use a fallback value of `true` for the rare platform that does not
support intersection observers. This is just to keep things from
crashing. Perf will be noticeably worse there, but that should be a
small subset of devices. We could polyfill if we want.
* Updated all the section components to use the new hook and switch
`enabled` based on visibility, also changed their rendering to be
switched based on `inView`.
* We render skeletons on initial load for all sections to make sure they
have a defined height from first render. Otherwise we will end up
triggering the intersection observer for all sections immediately.
* The `ExploreSection` was just doing some switching logic on Tile
types. I moved that up into the components to make this new style of viz
checking easier and deleted `ExploreSection`.
* Implemented loading skeletons for all sections and updated as many as
I could to avoid using the underlying "real" element with a loading
state. This reduces the work needed on initial render since we don't
have to process a bunch of hooks with empty/invalid values.
* Updated all skeletons on this page to use `noShimmer` to reduce the
load
* Fixed tile types between mobile/desktop in the various sections. It's
grid/small variant for the carousels, but large for the feeling lucky
tile. Fun fact, on mobile we use `large` stats in the profile tracks
lineup. So I'm matching that here.

### How Has This Been Tested?
The perf profiling time improved a lot and it's noticeable just in
interacting with the page.




https://github.com/user-attachments/assets/86f06c84-4c1a-47ab-b9a6-29735a65f4c2