[PE-6570] Hide user contextual content when logged out (#12695)
### Description
Skip rendering the sections that are specific to the current logged in
user if there isn't one. Also decided to remove the inView checks for
quick search and moods since they don't fetch data and the intersection
observer was causing them to flash nothing on their first render.

### How Has This Been Tested?
Local web/mobile web clients against staging, both logged in an
incognito.