Fix trending playlists playback on web (#12712)
### Description
Lineups
Are.
Gnarly.

I made a mistake when implementing this originally and we weren't
letting the uid from the tracks inside the collection tiles to flow
through to the actions. This was resulting in most of the playlists
sending the wrong UID for playback, which would get ignored by the queue
sagas and it would just call `next` can begin playing at the top of the
trending playlists lineup.

### How Has This Been Tested?
Explore page on web, just click around in various trending playlists
(including play buttons on track item rows) and make sure we correctly
change the queued item. Also tested next/previous across collection
boundaries and it matches the behavior from the trending playlists page.