Web Search/Explore QA fixes (#12717)
### Description
* Fix mood buttons doing nothing when clicked from 'All' category. I
updated this to match mobile, where we set the category to 'tracks' if
you are coming from the 'All' filter. It's invalid to search across all
items with a mood, since profiles don't support it
* Fix empty `query=` when loading the explore page or clearing your
search query
* Highlight 'Explore' nav item when the route begins with `/search`. I
added support to `LeftNavLink` for passing additional paths to match,
since it was set up to only allow matching on the `to` prop.
* Fix 'Active Discussions' loading dozens of items despite passing a
limit. I missed wiring up the `pageSize` prop a few PRs ago 😅
* Fix issue where Feeling Lucky track would show previous artwork for a
bit when refreshing.

### How Has This Been Tested?
A lil' click here, a lil' click there.