Parallelize and cache user pubkeys in chat create (#12719)
### Description
In Chats SDK, Add a cache for user pubkeys and parallelize fetch
requests for the pubkeys.

### How Has This Been Tested?

Tested on local web stage
Timing was fairly inconsistent. Before this change I was observing chat
creates taking ~4.5 seconds.

After the change I observed instances as low as 3.5s and as high as 6s..
guessing network instability. But I did notice that we saved a call
to`/pubkey` for the current user, so this is probably still worth
shipping?