Add a little more logging to help debug change email flow (#12735)
### Description
Was trying to debug a case of failing to change email due to artifacts
already existing, but we don't differentiate between the case where the
old and new keys match and we don't log out the key used to find the
already existing artifacts. Adding a new conditional and logging the new
lookupKey to help with debugging in those two cases.

### How Has This Been Tested?
N/A, will make sure e2e passes before deploying this.