Remove revert blocks test (#12770)
We no longer revert blocks anyway. Remove this test as it was blocking
staging deploy

Unsure why it started failing, but I suspect it's a broken test not
broken functionality. I think DDL changed and maybe we're not writing
the first block into the db or something like that, but honestly not
sure.