[QA-2255] Fix mobile lineup & playback bugs (continues playing same track after playing a new lineup page) (#12814)
### Description

Several key fixes for some hard to find bugs in here:
- Lineup saga is clearing the queue even when on a queue append, this
was one way we were causing our bug but not the root cause. Removing
- Tweak the `handleQueueChange` to properly wait for tanQuery data
before attempting to queue anything into RNTP; this was the root cause
of the bugs
- Add try/catch to `makeTrackData` - this was a very nasty side effect
of our bug where there was an error thrown inside this detached `async`
fn but never gets thrown anywhere so it just silently fails. The issue
with it silently failing is that the `await enqueueJobsRef` is stuck
infinitely
- Also fixed the issue where analytics was showing `Playback: Play
undefined` (unrelated to the playback bugs but fixed lol)

### How Has This Been Tested?

ios:stage simulator