Fix nginx config for comms websockets in bridge server (#12828)
### Description
Since local dev uses nginx and it doesn't forward the websocket headers
by default, we need this block (adapted from a similar entry for the DN
comms server) in order for websocket connection upgrades to succeed.

### How Has This Been Tested?
Run a bridge server that has comms ws support locally and make sure that
socket connections from local client succeed with a 101 status code and
not a 426.