[PE-6552] Fix buy/sell modal close issue (#12949)
### Description

When you close the modal after a success purchase, it flashes the
initial state before close. Modal has a prop for onClosed, so pass a
prop a reset fn setter to the inner component and use that on onClosed
instead of on success alone.

Also prevents dismissing the modal w/ a click outside when you are on
the confirm screen.

### How Has This Been Tested?

_Please describe the tests that you ran to verify your changes. Provide
repro instructions & any configuration._

Purchased, bailed out of flow, clicked outside of flow to bail out, etc.