Improve markdown styles for proposal descriptions (#12952)
### Description

Governance Proposals on the protocol dashboard are rendered with
markdown, but unfortunately there are site-wide styles that clash with
the appearance of the description.

This PR:
* updates react-markdown and remark-gfm to the latest versions (just for
the dashboard)
* properly renders the style and margins for bulleted lists
* adds spacing between paragraphs
* adds soft line-breaks for single newlines

### How Has This Been Tested?

Viewed currently open proposal using locally built instance and verified
the appearance. Also checked console to verify no errors relating to the
upgraded markdown components appeared.