[PE-6898] Improve create coin success (#12963)
### Description

Migrated the create-coin-success modal to it's own globally aware modal
since we need to navigate from the launchpad page to coin details page.

One question for @DejayJD, do i need to pass all that info along to the
modal, or is mint/ticker sufficient? i should be able to just request it
right?