[PE-6925] Artist coin table vol 2 digits (#12973)
### Description
We had a few different versions of `formatCount` in a few diff places.
Tried to centralize it into a location in `common` that we can use in
`createCoinMetrics`. Deleted an old version that seemed not in-use as
well.

left the one in harmony tho cuz it's used in NotificationCount.. sigh
duplicated code

### How Has This Been Tested?

<img width="869" height="765" alt="Screenshot 2025-09-24 at 1 30 36 PM"
src="https://github.com/user-attachments/assets/37ae6f73-85f1-4226-8dde-190155982ddd"
/>