[PE-6875] First buy retry UX flow (#12983)
### Description

![2025-09-24 15 14
23](https://github.com/user-attachments/assets/5c4df982-94b3-4416-8b37-c614c3d6550c)
(gif is showing a hardcoded version of what this scenario looks like)

NOTE: this is specifically a UX flow for the user to retry the first buy
TX, not to be confused with automated retries.

Add a retry UX flow for the possible event that the first buy fails
(reasons for this could be: solana issues, user cancelled tx)

Changes:
* Adds UI tweaks (toasts/error messaging) that notify the user when the
coin was created but the first buy failed (Sammie provided copy)
* Adds a retry swap flow using a different TX than the original first
buy, just through a jupiter swap. The reasons for making a completely
new swap TX is:
* There's not much special about the first buy swap TX that meteora
generates for us, its a standard regular pool buy swap TX
* Since we're putting them back on the form UI, making a new swap means
they can change the form inputs if they want (reusing the original TX
would mean locking the inputs). This could be especially relevant if the
reason they cancelled the first buy was because they wanted to change
the amount
* If there was an issue with the first buy TX (unlikely), generating a
new TX could be the solution

### How Has This Been Tested?

web:prod - made a new coin, cancelled the first buy, went back through
this flow & successfully purchased again