Remove TX skip consensus behavior (#13085)
### Description
This removes the redis/peer-based checking for "consensus" on failed tx
parsing in favor of just immediately skipping, logging, and inserting a
DB record. This code path should only get exercised for uncaught errors
(validation errors are automatically ignored).

Also removes the endpoints that are used to get consensus (the
`/indexing` paths) and undoes my nginx change to expose those URLs. We
can directly query the DB to get skipped transactions.

Plan is to add a monitor for the specific error that we log when this
happens and we can investigate if it happens again.

### How Has This Been Tested?
Integration tests and local stack